+++ src/Forms/components/FormEditor/FormEditorModal.js
@@ -1,45 +1,15 @@
-import React, { useState, useEffect} from "react";
+import React from "react";
 import { Button, Dialog, DialogContent, DialogTitle, Grid, Typography } from "@material-ui/core";
-// import { Edit } from "@material-ui/icons";
 import { PropTypes } from 'prop-types';
 import { Light as SyntaxHighlighter } from "react-syntax-highlighter";
 import docco from "react-syntax-highlighter/dist/esm/styles/hljs/stackoverflow-light"
-import { useAuth0 } from "../../../Auth/react-auth0-spa";
 import dark from "react-syntax-highlighter/dist/esm/styles/hljs/stackoverflow-dark";
-import { callAzureFunction } from "../../../utils/SharedFunctions" 
 
 import EditableField from "./EditableField";
 
 const FormEditorModal = ( props ) => {
-    let { isDarkTheme, modalOpen, toggleModalOpen, slimRecord } = props
-
-    const [editableList, setEditableList] = useState("")
-    const [loading, setLoading] = useState(true)
-    const { getTokenSilently } = useAuth0();
-    // console.log(modalOpen)
-
-    const fetchEditableList = async () => {
-        const data = { version: slimRecord.__version__};
-        const res = await callAzureFunction(data, "EditableList", getTokenSilently);
-        console.log(res.jsonResponse);
-        setEditableList(JSON.parse(res.jsonResponse));
-        setLoading(false)
-        console.log(editableList);
-    }
-
-    useEffect(() => {
-        fetchEditableList()
-    }, [])
+    let { isDarkTheme, modalOpen, toggleModalOpen, slimRecord, editableList } = props
     
-
-    console.log(props);
-    
-    // console.log(editableList)
-    // let jsonList = JSON.parse(editableList.jsonResponse)
-    // let jsonList = []
-
-    
-
     return (
         typeof(modalOpen) === "boolean" && modalOpen ? 
             <Dialog 
@@ -68,8 +38,7 @@ const FormEditorModal = ( props ) => {
                                 <Typography variant="h6">Editable Entries</Typography>
                             </Grid>
                             <Grid item>
-                                {!loading && editableList.map((entry, index) => {
-                                    // console.log(entry);
+                                {editableList.map((entry, index) => {
                                     return <EditableField field={entry} data={slimRecord[entry]} key={index} />
                                 })}
                             </Grid>                            
@@ -112,7 +81,7 @@ FormEditorModal.propTypes = {
     modalOpen: PropTypes.bool, 
     toggleModalOpen: PropTypes.function,
     editableList: PropTypes.object,
-    slimRecord: PropTypes.object
+    slimRecord: PropTypes.object,
 }
 
 export default FormEditorModal
